refactor(LC015): split missing-orderby analysis helpers
- separate upstream and downstream ordering analysis
- keep analyzer behavior unchanged while reducing local complexity
